Implements logout functionality
Adds a logout button to the header.
Implements the logout function to clear tokens and redirect the user to the authentication page.

The logout function calls the backend logout endpoint.
On successfull backend logout or in case of an error, tokens are cleared, and the user is redirected to the auth page using `window.location.href` to avoid route guards.

Updates the auth view to include a logo.