Adds script to add test users to the database
Introduces a new script to populate the database with predefined test users for development and testing purposes.
The script automatically adds these users to the 'managelog' company, creating the company if it does not exist.

Also, implements project participants selection feature in the frontend.
It allows users to assign existing users to projects.